+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Upgraded to `datafusion` `v37.1.0`
 - Split the Flow system crates
+### Fixed
+- `kamu system diagnose` command crashes when executed outside the workspace directory
 
 ## [0.176.3] - 2024-04-18
 ### Changed